Move MotionEntity and Enumerations into new group base inside entities
