+++ ios/RNSketchCanvas/RNSketchCanvas/entities/TriangleEntity.m
@@ -7,7 +7,7 @@
 //
 
 #import <Foundation/Foundation.h>
-#import "MotionEntity.h"
+#import "base/MotionEntity.h"
 #import "TriangleEntity.h"
 
 @implementation TriangleEntity
