+++ ios/RNSketchCanvas/RNSketchCanvas/entities/base/Enumerations.h
