+++ ios/RNSketchCanvas/RNSketchCanvas/entities/base/MotionEntity.h
