+++ ios/RNSketchCanvas/RNSketchCanvas/entities/base/MotionEntity.m
